requirements.txt
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-timezone-field==2.1
djangorestframework==3.7.3
django-rest-swagger==2.1.2
